client-api: Move Capabilities into discovery::get_capabilities::v3 module
And merge the files, for consistency with other endpoints.
